Remove rl-loadout-lib and replace Material-UI with the updated library
rl-loadout-lib is mostly useless right now and is just bloat with its dependencies. The material-ui library is also renamed to @mui and is in a new location. Requires some minor changes to imports and refactoring of some components. Seems to work okay though there are some visual issues with the new library in /docs.